+++ test/audiovideocontroller/DefaultAudioVideoController.test.ts
@@ -1596,6 +1596,10 @@ describe('DefaultAudioVideoController', () => {
         getMidForStreamId(streamId: number): string | undefined {
           return streamId.toString();
         }
+
+        hasVideoInput(): boolean {
+          return true;
+        }
       }
 
       // @ts-ignore
@@ -1757,6 +1761,19 @@ describe('DefaultAudioVideoController', () => {
           };
         }
       }
+
+      class TestTransceiverController extends DefaultTransceiverController {
+        getMidForStreamId(streamId: number): string | undefined {
+          return streamId.toString();
+        }
+
+        hasVideoInput(): boolean {
+          return true;
+        }
+      }
+
+      // @ts-ignore
+      audioVideoController.meetingSessionContext.transceiverController = new TestTransceiverController();
       // @ts-ignore
       audioVideoController.meetingSessionContext.transceiverController.localVideoTransceiver().sender = new MockRTCRtpSender();
       // @ts-ignore
@@ -1794,8 +1811,18 @@ describe('DefaultAudioVideoController', () => {
 
       await start();
 
+      class TestTransceiverController extends DefaultTransceiverController {
+        getMidForStreamId(streamId: number): string | undefined {
+          return streamId.toString();
+        }
+
+        hasVideoInput(): boolean {
+          return true;
+        }
+      }
+
       // @ts-ignore
-      audioVideoController.meetingSessionContext.transceiverController.localVideoTransceiver().sender = new RTCRtpSender();
+      audioVideoController.meetingSessionContext.transceiverController = new TestTransceiverController();
       // @ts-ignore
       audioVideoController.meetingSessionContext.lastVideosToReceive = new DefaultVideoStreamIdSet([
         1,
@@ -1812,6 +1839,53 @@ describe('DefaultAudioVideoController', () => {
       // Slightly awkward logger check since subscribe steps are asynchronous and hard to capture
       expect(loggerSpy.calledWith(sinon.match('Update request does not require resubscribe'))).to.be
         .false;
+    });
+
+    it('will skip renegotiation if we are updating simulcast layer but not sending', async () => {
+      const logger = new NoOpDebugLogger();
+      const loggerSpy = sinon.spy(logger, 'info');
+      configuration.enableSimulcastForUnifiedPlanChromiumBasedBrowsers = true;
+      domMockBehavior.browserName = 'chrome';
+      domMockBuilder = new DOMMockBuilder(domMockBehavior);
+
+      audioVideoController = new DefaultAudioVideoController(
+        configuration,
+        logger,
+        webSocketAdapter,
+        new NoOpMediaStreamBroker(),
+        reconnectController
+      );
+
+      class TestTransceiverController extends DefaultTransceiverController {
+        getMidForStreamId(streamId: number): string | undefined {
+          return streamId.toString();
+        }
+
+        hasVideoInput(): boolean {
+          return false;
+        }
+      }
+
+      // @ts-ignore
+      audioVideoController.meetingSessionContext.transceiverController = new TestTransceiverController();
+      // @ts-ignore
+      audioVideoController.meetingSessionContext.lastVideosToReceive = new DefaultVideoStreamIdSet([
+        1,
+      ]);
+      // @ts-ignore
+      audioVideoController.meetingSessionContext.videosToReceive = new DefaultVideoStreamIdSet([1]);
+
+      await start();
+
+      // @ts-ignore
+      audioVideoController.mayNeedRenegotiationForSimulcastLayerChange = true;
+      audioVideoController.update({ needsRenegotiation: false });
+
+      await stop();
+
+      // Slightly awkward logger check since subscribe steps are asynchronous and hard to capture
+      expect(loggerSpy.calledWith(sinon.match('Update request does not require resubscribe'))).to.be
+        .true;
     });
 
     it('will skip renegotiation if we are only completing simulcast stream switches', async () => {